pyproject.toml
[project]
name = "lms-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "django==6.0.1",
    "channels>=4.2",
    "daphne>=4.0",
    "djangorestframework>=3.15",
    "drf-yasg>=1.21",
    "djangorestframework-simplejwt>=5.4",
    "django-filter>=24.0",
    "django-cors-headers>=4.0",
    "stripe>=14.4.1",
    "psycopg[binary]>=3.2",
    "openpyxl>=3.1",
]
